+++ server/apiCalls/koppsApi.js
@@ -3,7 +3,7 @@
 const log = require('@kth/log')
 const redis = require('kth-node-redis')
 const connections = require('@kth/api-call').Connections
-const config = require('./configuration').server
+const config = require('../configuration').server
 
 const koppsOpts = {
   log,
